// vuejs repo/src/api/generated/models/RoleDTOPaginationItemsApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RoleDTOPaginationItems } from './RoleDTOPaginationItems'
export type RoleDTOPaginationItemsApiResponse = {
  requestId?: string
  isSuccess?: boolean
  message?: string
  data?: RoleDTOPaginationItems
}